# Conflicts:
#	src/main/java/it/polimi/ingsw/client/ModelClient.java
#	src/main/java/it/polimi/ingsw/controller/Controller.java
#	src/main/java/it/polimi/ingsw/server/Match.java

+++ src/main/java/it/polimi/ingsw/model/personalBoard/cardManager/CardManager.java
@@ -78,7 +78,9 @@ public class CardManager extends GameMasterObservable implements Observable<Card
         notifyGameMasterObserver(GameMasterObserver::discardLeader);
         notifyAllObservers(x -> x.leaderDiscard(leaderIndex));
 
-        deleteAllCreationEffects(leaderToDiscard);
+        if (leaderToDiscard.isActive()){
+            deleteAllCreationEffects(leaderToDiscard);
+        }
     }
 
     public void discardLeaderSetUp(int leaderIndex) throws IndexOutOfBoundsException{
